# server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WOSP API Documentation",
        "description": "WOSP API in Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "3.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "WOSP API Server"
        }
    ],
    "paths": {
...
        "/charityBoxes": {
            "get": {
                "tags": [
                    "CharityBoxes"
                ],
                "summary": "Get list of charity boxes",
                "description": "Returns list of charity boxes",
                "operationId": "getCharityBoxList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CharityBoxResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "CharityBoxes"
                ],
                "summary": "Create charity box for volunteer (collector)",
                "description": "Returns information about realse charity box for volunteer (collector)",
                "operationId": "createCharityBox",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CollectorCharityBoxRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CharityBox"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/charityBoxes/{id}": {
            "get": {
                "tags": [
                    "CharityBoxes"
                ],
                "summary": "Get charity box information",
                "description": "Returns chrity box data",
                "operationId": "getCharityBoxById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Charity box id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CharityBox"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "CharityBoxes"
                ],
                "summary": "Update charity box",
                "description": "Returns information about modified charity box",
                "operationId": "updateCharityBox",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Charity box id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCharityBoxRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CharityBox"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/charityBoxes/unverified": {
            "get": {
                "tags": [
                    "CharityBoxes"
                ],
                "summary": "Get list of unverified charity boxes",
                "description": "Returns list of charity boxes",
                "operationId": "getUnverifiedList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CharityBoxResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/charityBoxes/verified": {
            "get": {
                "tags": [
                    "CharityBoxes"
                ],
                "summary": "Get list of verified charity boxes",
                "description": "Returns list of charity boxes",
                "operationId": "getVerifiedList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CharityBoxResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/charityBoxes/verified/{id}": {
            "post": {
                "tags": [
                    "CharityBoxes"
                ],
                "summary": "Verify Charity Box",
                "description": "Return message of operation result",
                "operationId": "postVerifyCharityBoxById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Charity box id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/charityBoxes/unverified/{id}": {
            "post": {
                "tags": [
                    "CharityBoxes"
                ],
                "summary": "Unverify Charity Box",
                "description": "Return message of operation result",
                "operationId": "postUnverifyCharityBoxById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Charity box id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/charityBoxes/{id}/startCounting": {
            "post": {
                "tags": [
                    "CharityBoxes"
                ],
                "summary": "Start counting Charity Box",
                "description": "Return charity box instance",
                "operationId": "stratCountngCharityBoxById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Charity box id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CharityBox"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/charityBoxes/{id}/finishCounting": {
            "post": {
                "tags": [
                    "CharityBoxes"
                ],
                "summary": "Finish counting Charity Box",
                "description": "Return charity box instance",
                "operationId": "finishCountngCharityBoxById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Charity box id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CharityBox"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CollectorCharityBoxRequest": {
                "title": "",
                "description": "",
                "required": [
                    "collector_identifier"
                ],
                "properties": {
                    "collector_identifier": {
                        "title": "collector_identifier",
                        "description": "Id of the collector identifier",
                        "type": "integer",
                        "format": "int64",
                        "example": "1"
                    }
                },
                "type": "object"
            },
...
            "CharityBoxResource": {
                "title": "Charity Box Resource",
                "description": "Charity Box resource",
                "properties": {
                    "data": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CharityBox"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "CharityBoxResource"
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "basic"
            }
        }
    },
...
}